safe
Sometimes, like this moment, I really don't like my framework, mostly
just because of some, maybe, tiny decificencies. Like this time, I
thought locale in Isabelle is beautiful as told by my teacher, but now I
found it is not. It replays every operation causing it is essentially
unusable for large-scale framework like mine. It starts to become slow
now.

I have proposed a better and also realistic solution for this,
axiomization. I know, if I really apply this, hundereds of questions
will come from paper reviewers. It is true, if we want it to be really
elegant, it must be an extension on HOL logic level. We need to extend
the type definition mechanism of Isabelle/HOL. For sure, it will spend a
paper to discuss this, to show its acceptable trust kernel. It is the
point I dislike the academic circle. It indeed limits me to do the right
thing, because I have to publish papers in order, so that we are able to
clarify everything.

The current work sequence is not correct for publishing papers. It is
not that terrible because we need to gather resources as fast as we
can. We are working for some bussinessial exploiters, eating my flesh.
We really need resources and because of this I must give a demo, no matter
how ugly it is, to demonstrate our value. It again blocks me to do the
right thing.

I really don't like this. I have to swallow the pain.