Fix bug in generic_variable_access.ML
A problem is, the variables in syntax translation are required having
the external names. However, over generic variable assignment mechanism
extracts that of internal names. It makes me feel really uncomfortable
that we have to convert them into the external name and than during the
following syntax parse they are parsed to the internal name back, to
something we cannot really control, while we already have the good,
precise internal term of what we want...