Introduce carrier set
It seems an explicit term of carrier set is still inevitable, so we
formalize a type class to fix one carrier set of the algebra
represented by the type. Later we also need the notion of (a variant of)
Abstract Domain that falls in the carrier, together with a deriver
deriving it. It shouldn't be hard once the carrier is fixed to each
type.