Give up φ-LPR based Premise Extraction (the previous commit)
After a shower-time thinking, I decide to give up the previous commit,
the φ-LPR based Premise Extraction. Yes, the φ-LPR based one is
powerful especially in extensionibility. However, it is too heavy and
slow. Obligation collection is considered as a small and quick step
in the reasoning. However, inside this small step it invokes several
sub-reasoning for its premises, where every sub-reasoning requires
initialization, fixing variables, and unfixing them later. The quantity
of those premises requiring extraction is not small.

By contrast, the original implementation is fine-tuned, elaboratedly
optimized, using rewrite rules with minimal resolution, based on fast
pattern match. Yes, sure, it is not that general and extensible, but still
can support any demand by ML code, though a bit cumbersome.

Actually I question, whether premises are so flexible that needs a such
powerful φ-LPR reasoning. Most of premises should be in a simple form
right? mainly in a single predicate. I cannot answer. I am not sure.

The best solution, now I have realized. Based on the original
implementation, we can in addition provides an 'extractor' falling back
to φ-LPR. That is we can have two mechanisms meantime. For the frequent
premises, the fine-tuned ML code is used, and for others, φ-LPR rules
can be used to extend quickly the system.