Stop here today and lett Tyops_Comm_I pass.
Several deficiencies have been fixed, but the deriving for the
commutativity between List and +_φ still fails, seemly missing certain
condition during the reasoning but I have no time to find it.
The abstract object side and itnermediate conditions really need
some simplification to reduce the exploded size of lambda terms and to
save human's eyes... so that maybe later I can find where the problem
is.