The module of extracting pure facts works
Now the general extraction of pure facts and the premise extraction are
uniformed.
Proof obligations in deriving φ-TA properties are augmented with pure
facts implifed in the inductive hypotheses.

Simplifier for the inferred properties has to be improved as the
infferred conditions now are full of redundent boolean conditions
extracted from the antecedents.
I am thinking using some φ-LPR reasoning to simplify the inferred rules.

The dervier for singular objective equivalence will be revoked, because
it lacks of a simple way to distinguish from the given annotations the
properties gonna to be derived normally or using the singular deriving,
as the reachability relation may not be the constant true (which was the
previous way to distinguish them).
Instead, I will introduce the notion of unital type as the singular case
of Identity_Element, from which the singular case of Object_Equiv will
be derived (though not compete).